+++ src/sll/api/log.c
@@ -4,6 +4,7 @@
 #include <sll/common.h>
 #include <sll/location.h>
 #include <sll/log.h>
+#include <sll/scheduler.h>
 #include <sll/types.h>
 #include <sll/vm.h>
 
@@ -12,7 +13,7 @@
 __API_FUNC(log_log){
 	sll_string_index_t fp;
 	sll_string_index_t fn;
-	sll_file_offset_t ln=sll_get_location(sll_current_runtime_data->a_dt,sll_current_instruction_index,&fp,&fn);
+	sll_file_offset_t ln=sll_get_location(sll_current_runtime_data->a_dt,sll_thread_get_instruction_index(SLL_UNKNOWN_THREAD_INDEX),&fp,&fn);
 	sll_string_t s;
 	sll_api_string_convert(a,ac,&s);
 	sll_log_raw((sll_current_runtime_data->a_dt->st.dt+fp)->v,(fn==SLL_MAX_STRING_INDEX?SLL_CHAR("@code"):(sll_current_runtime_data->a_dt->st.dt+fn)->v),ln,0,&s);